requirements.txt
requests
python-dotenv
openai
schedule
